refactor: remove `constants` module and update related imports
- delete `constants` module and its documentation
- replace imports of `constants` with `status-code` in `http-client`
- adjust imports in `access-control` to match new structure
- update `tsdown` and `package.json` for improved modularity and maintainability
- clean up `pnpm-lock.yaml` by removing unused dependencies
